CustomerDAO and List Customers Feature

+++ src/main/java/com/bookstore/service/backend/CategoryServices.java
@@ -11,21 +11,20 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import com.bookstore.dao.BookDAO;
 import com.bookstore.dao.CategoryDAO;
 import com.bookstore.dao.UserDAO;
 import com.bookstore.entity.Category;
 
 public class CategoryServices {
-	private EntityManager entityManager;
 	private CategoryDAO categoryDAO;
 	private HttpServletRequest request;
 	private HttpServletResponse response;
 	
-	public CategoryServices(EntityManager entityManager,HttpServletRequest request, HttpServletResponse response) {
+	public CategoryServices(HttpServletRequest request, HttpServletResponse response) {
 		this.request=request;
 		this.response=response;
-		this.entityManager=entityManager;
-		categoryDAO=new CategoryDAO(entityManager);
+		categoryDAO=new CategoryDAO();
 	}
 
 	public void listCategory(String message) throws ServletException, IOException {
@@ -90,8 +89,17 @@ public class CategoryServices {
 
 	public void deleteCategory() throws ServletException, IOException {
 		int categoryId=Integer.parseInt(request.getParameter("id"));
-		categoryDAO.delete(categoryId);
-		String message="The category with ID "+categoryId+" has been removed successfully!";
+		BookDAO bookDAO=new BookDAO();
+		long numberOfBooks=bookDAO.countByCategory(categoryId);
+		String message;
+		if (numberOfBooks > 0) {
+			message="Could not delete the category because it contains (%d) books .";
+			message=String.format(message, numberOfBooks);
+		}else {
+			categoryDAO.delete(categoryId);
+			message="The category with ID "+categoryId+" has been removed successfully!";
+		}
+		
 		listCategory(message);
 		
 	}
